% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wienR_helpers.R
\name{binary operator}
\alias{binary operator}
\alias{\%||\%}
\title{Helper function for null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Condition 1}

\item{y}{Condition 2}
}
\description{
If x is null, y. Otherwise, x.
}
\keyword{internal}
