% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workspace.R
\name{clean_workspace}
\alias{clean_workspace}
\title{Clear the console and plot window, and remove objects from the environment}
\usage{
clean_workspace()
}
\value{
confirmation that the environment has been cleaned
}
\description{
Clear the console and plot window, and remove objects from the environment
}
\examples{
clean_workspace()
}
