% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.cognition.import.R
\name{csv.cognition.import}
\alias{csv.cognition.import}
\title{Import CSV file data downloaded from jsPsych and Cognition.run. This function
assumes data files are stored in a folder named "Data" in the working directory}
\usage{
csv.cognition.import(select = tidyselect::everything())
}
\arguments{
\item{select}{A vector of column names or numbers to keep, drop the rest.
Select may specify types too in the same way as colClasses; i.e., a vector of
colname=type pairs, or a list of type=col(s) pairs. In all forms of select,
the order that the columns are specified determines the order of the columns
in the result.}
}
\value{
A data frame, a matrix-like structure whose columns may be of
differing types (numeric, logical, factor and character and so on).
}
\description{
Import CSV file data downloaded from jsPsych and Cognition.run. This function
assumes data files are stored in a folder named "Data" in the working directory
}
\examples{
csv.cognition.import(select = c(phase = 'character', trial_type = "character",
trial_index = "numeric", time_elapsed = "numeric", run_id = "numeric",
condition = "factor", recorded_at = "character", rt = "numeric",
stimulus = "character", response = "character"))
}
