% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_short.R
\name{date_short}
\alias{date_short}
\title{Date Strings}
\usage{
date_short(time = FALSE)
}
\arguments{
\item{time}{Logical. Should time be included in the short date string?
Defaults to FALSE.}
}
\value{
A string corresponding to the current system date
}
\description{
A function to output simple formatted date strings
}
\examples{
# Output the current date only
date_short()

# Output the current date and time
date_short(time = TRUE)
}
