% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{files_info}
\alias{files_info}
\title{File Information Compiler}
\usage{
files_info(
  path = system.file("extdata", package = "pcjtools"),
  extension = "csv",
  ...
)
}
\arguments{
\item{path}{A filepath string that names a \emph{directory} with files to read.}

\item{extension}{A string naming the extension of the files to import.}

\item{...}{Optional logical arguments of the following to pass to
\code{list.files()}: \code{'all.files'}, \code{'recursive'}, \code{'ignore.case'},
\code{'include.dirs'}. All default to \code{FALSE}.}
}
\value{
A data.table containing file information from the file(s) found in
the directory specified at \code{'path'}. Including:
\itemize{
\item full filepath
\item basename
\item extension
\item the last date the file was modified
\item the age in days since the file was last modified.
}

This information can be stored in a vector and is intended to be passed to
\code{import_data()} to read in a list of data files to a single \code{data.table}
object.
}
\description{
A function to gather information about data files located at \code{'path'}. Only
files of the type specified in \code{'extension'} will be returned.
}
\examples{
files_info()
}
