% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj_helpers.R
\name{font_install}
\alias{font_install}
\title{Font Installer}
\usage{
font_install(font)
}
\arguments{
\item{font}{Provide a string naming the font to install if not already
installed.}
}
\value{
Invisibly installs fonts.
}
\description{
A helper function to install the Atkinson Hyperlegible font if it isn't
already installed. Atkinson Hyperlegible is the default font used by
\code{\link[=theme_pcj]{theme_pcj()}}.
}
\keyword{internal}
