% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{A function to import jsPsych CSV data files downloaded from cognition.run}
\usage{
import_csv(file.dir = NULL, file.type = NULL)
}
\arguments{
\item{file.dir}{a string. The the filepath of the directory containing the files to be read in.}

\item{file.type}{a string. The name of a type of file extension (e.g., "xlsx", "csv", etc.)}
}
\value{
an object containing the data from the files in the specified directory.
}
\description{
A function to import jsPsych CSV data files downloaded from cognition.run
}
\examples{
import_csv(file.type = "csv")
}
