% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{install_atkinson}
\alias{install_atkinson}
\title{A function to install the Atkinson Hyperlegible font required for the
theme_pcj}
\usage{
install_atkinson()
}
\description{
A function to install the Atkinson Hyperlegible font required for the
theme_pcj
}
\examples{
install_atkinson()
}
