% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{load_atkinson}
\alias{load_atkinson}
\title{A function to download and import the Atkinson Hyperlegible font required for
the theme_pcj}
\usage{
load_atkinson(overwrite = FALSE, os_fonts = "win")
}
\arguments{
\item{overwrite}{logical. If font files are downloaded to the system folder, should
existing Atkinson font files be overwritten?}

\item{os_fonts}{a string, including "all", "pdf", "postscript" and "win",
indicating which fonts should be loaded into RStudio.}
}
\description{
A function to download and import the Atkinson Hyperlegible font required for
the theme_pcj
}
\examples{
\dontrun{
load_atkinson(overwrite = FALSE, os_fonts = "win")
}
}
