% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load a list of packages}
\usage{
load_packages(x)
}
\arguments{
\item{x}{A character list of packages to load}
}
\value{
A message confirming the specified packages have been loaded, or an error.
}
\description{
Load a list of packages
}
\examples{
\donttest{
packages <- c("dplyr", "ggplot2")

load_packages(x = packages)
}
}
