% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nll}
\alias{nll}
\title{Find The Negative Log Likelihood of a Parameter}
\usage{
nll(
  par,
  rt,
  response,
  bound_index,
  drift_index,
  resid_index,
  sv_index = NULL,
  sw_index = NULL,
  st0_index = NULL,
  ...
)
}
\arguments{
\item{par}{Start}

\item{rt}{Start}

\item{response}{Start}

\item{bound_index}{Start}

\item{drift_index}{Start}

\item{resid_index}{Start}

\item{sv_index}{Start}

\item{sw_index}{Start}

\item{st0_index}{Start}

\item{...}{Optional arguments to pass to \code{'WienerPDF()'}.}
}
\value{
Something
}
\description{
A helper function used by \verb{'fit_wienr()} to find the negative log likelihood
of a set of parameters
}
\examples{
\dontrun{
nll()
}
}
