% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Title}
\usage{
outliers(x, columns, sd_threshold = 2)
}
\arguments{
\item{x}{A dataframe object containing columns and rows of data.}

\item{columns}{A list of column names outliers will be calculated for.}

\item{sd_threshold}{The number of standard deviations from the mean an observation must be to be considered an outlier.}
}
\value{
Returns a list of the outlier values from the specified columns of a dataframe.
}
\description{
Title
}
\examples{
\dontrun{
outliers(x = dataframe, columns = c("Var1", "Var2"), sd_threshold = 2)
}
}
