% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{pcj_aesthetics}
\alias{pcj_aesthetics}
\title{Graph aesthetics to maximize legibility of text on graphs printed for academic posters}
\usage{
pcj_aesthetics(base_size = 12, dark_text = "#000000", ...)
}
\arguments{
\item{base_size}{an integer, the size the font in a plot should default to.}

\item{dark_text}{a quoted hex string, sets the color of the darkest text in a plot.}

\item{...}{options to be passed to function theme().}
}
\value{
a plot configured with aesthetic settings specified by options set by this function
}
\description{
Graph aesthetics to maximize legibility of text on graphs printed for academic posters
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
ggplot2::geom_point() +
pcj_aesthetics()
}
