% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{pcj_graph_palettes}
\alias{pcj_graph_palettes}
\title{A suite of functions to customize plot aesthetics}
\usage{
pcj_graph_palettes(
  palette = "default",
  continuous = FALSE,
  .colors = pcj_colors,
  .palettes = pcj_palettes
)
}
\arguments{
\item{palette}{a string, including: "default", "neg_to_pos", "mono_printing",
"mono_blue", "mono_red", "mono_yellow", or "ualbany".}

\item{continuous}{logical. Whether the scale of the data is discrete or continuous.}

\item{.colors}{loads colors defined in pcj_colors.}

\item{.palettes}{loads palettes defined in pcj_palettes.}
}
\value{
a plot object with the specified aesthetics rendered.
}
\description{
A suite of functions to customize plot aesthetics
}
\examples{
pcj_graph_palettes(palette = "default")
pcj_graph_palettes(palette = "mono_printing")
}
