% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{read_file}
\alias{read_file}
\title{a helper function used by 'read_file_list()' to read individual data files}
\usage{
read_file(x)
}
\arguments{
\item{x}{a string with file name paths}
}
\value{
a data.table object containing data from a file supplied to
'read_data_list()'
}
\description{
a helper function used by 'read_file_list()' to read individual data files
}
\examples{
\dontrun{
read_file(x)
}
}
\keyword{internal}
