% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{read_file_list}
\alias{read_file_list}
\title{A Reader for Lists of Files}
\usage{
read_file_list(files)
}
\arguments{
\item{files}{A string or an object containing filepath string(s).}
}
\value{
A \code{data.table} object with concatenated data from all named files
supplied in \code{'file_list'}
}
\description{
A wrapper to read in multiple files passed from the \code{'read_file()'} helper
function.
}
\examples{
info <- files_info()
data <- import_data(x = info$filepath)
}
\keyword{internal}
