% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_cols.R
\name{recode_cols}
\alias{recode_cols}
\title{An efficient variable recoder to change column types in a datatable}
\usage{
recode_cols(dt, col_types, .verbose = FALSE)
}
\arguments{
\item{dt}{A datatable with columns you wish to recode.}

\item{col_types}{A named list of columns to recode with how they should be
recoded in the style of col_name = "col_type".}

\item{.verbose}{Logical. Should results of recoding be reported?}
}
\value{
Invisibly returns the new datatable with updated columns
}
\description{
An efficient variable recoder to change column types in a datatable
}
\examples{
data <- data.table::data.table(
  name = rep(c("Alice", "Bob", "Charlie"), length.out = 10),
  age = as.character(rep(25:30, length.out = 10)),
  score = as.character(runif(10, 70, 100))
  )

recode_cols(data, list(
  age = "integer",
  score = "numeric"
  ))
}
