% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{The delimiter in x to split by.}
}
\value{
A character vector whose length will be the number of words from
the delimited work list in vector x.
}
\description{
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
