% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{theme_pcj}
\alias{theme_pcj}
\title{a function to create and save consistent plots}
\usage{
theme_pcj(
  plot,
  base_size = 12,
  dark_text = "#000000",
  font = "Atkinson Hyperlegible",
  palette = "default",
  continuous = FALSE,
  plot_text = c(title = "title", subtitle = "subtitle", xlab = "xlab", ylab = "ylab",
    caption = paste0("Rendered:", format(Sys.time(), "\%Y\%m\%d, \%H:\%M"))),
  alt_text = TRUE,
  save_path = NULL,
  ...
)
}
\arguments{
\item{plot}{a ggplot object to be formatted}

\item{base_size}{an integer. The graph's default font size.}

\item{dark_text}{a quoted hex string. Sets the color of the darkest text in a
plot. All text is based on shades of the specified hex code.}

\item{font}{a quoted name of the font you would like the plot's text to be
printed in}

\item{palette}{a string. The name of the palette to be mapped to a variable,
including: "default", "neg_to_pos", "mono_printing", "mono_blue", "mono_red",
"mono_yellow", or "ualbany".}

\item{continuous}{logical. Is the variable continuous or discrete?}

\item{plot_text}{a named list of plot features and their text label}

\item{alt_text}{logical. Should a subtitle and caption be generated for the
plot?}

\item{save_path}{a string. The filename path where the plot should be saved
directory}

\item{...}{additional arguments passed to \code{theme_pcj_aesthetics()} or
\code{ggsave()} functions}
}
\value{
a plot object with the specified aesthetics
}
\description{
a function to create and save consistent plots
}
\examples{
\dontrun{
theme_pcj(plot = plot)
}
}
