% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{theme_pcj_palettes}
\alias{theme_pcj_palettes}
\title{a helper function defining the palettes accessible to \code{theme_pcj()}}
\usage{
theme_pcj_palettes(
  palette = "default",
  continuous = FALSE,
  .colors = .colors,
  .palettes = .palettes
)
}
\arguments{
\item{palette}{a string. The name of the palette to be mapped to a variable.}

\item{continuous}{logical. Is the variable continuous or discrete?}

\item{.colors}{loads the colors defined within pcj_colors.}

\item{.palettes}{loads the palettes defined within pcj_palettes.}
}
\value{
a plot object with the specified aesthetics rendered
}
\description{
a helper function defining the palettes accessible to \code{theme_pcj()}
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
ggplot2::geom_point() +
theme_pcj_palettes()
}
