% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{theme_pcj_text}
\alias{theme_pcj_text}
\title{a function to apply text to a plot}
\usage{
theme_pcj_text(
  plot_text = c("title", "subtitle", "xlab", "ylab", paste0("Rendered:",
    format(Sys.time(), "\%Y\%m\%d, \%H:\%M"))),
  alt_text = FALSE
)
}
\arguments{
\item{plot_text}{a named list of plot features and their text label}

\item{alt_text}{logical. Should a subtitle and caption be generated for the
plot?}
}
\value{
a ggplot object with the rendered text specified in the function
call
}
\description{
a function to apply text to a plot
}
\examples{
\dontrun{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
  ggplot2::geom_point() +
  theme_pcj_text()
  }
}
\keyword{internal}
