% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wienR_helpers.R
\name{binary operator}
\alias{binary operator}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check.}

\item{y}{Default value if \code{x} is \code{NULL}.}
}
\value{
Either \code{x} (if not \code{NULL}) or \code{y}.
}
\description{
Returns \code{y} if \code{x} is \code{NULL}, otherwise returns \code{x}.
}
\keyword{internal}
