% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workspace.R
\name{clean_workspace}
\alias{clean_workspace}
\title{Clear the console and plot window, and remove objects from the environment}
\usage{
clean_workspace(ask = TRUE)
}
\arguments{
\item{ask}{Logical; should the function ask for confirmation before clearing?
Defaults to TRUE for safety.}
}
\value{
confirmation that the environment has been cleaned (invisibly)
}
\description{
Clear the console and plot window, and remove objects from the environment
}
\examples{
\donttest{
# Clean workspace without asking (use with caution!)
clean_workspace(ask = FALSE)
}
}
