% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workspace.R
\name{clean_workspace}
\alias{clean_workspace}
\title{Work Space Cleaner}
\usage{
clean_workspace(confirm = TRUE)
}
\arguments{
\item{confirm}{Logical (default is TRUE). Should the function ask for
confirmation before clearing?}
}
\value{
Confirmation that the environment has been cleaned.
}
\description{
Clear the console, plot window, and remove objects from the global
environment.
}
\examples{
\donttest{
# Execute without asking (use with caution)
clean_workspace(confirm = FALSE)
}
}
\seealso{
\code{\link[=graphics.off]{graphics.off()}} and \code{\link[=rm]{rm()}}, which clean_workspace serves
as a wrapper for.
}
