% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_mod_helper.R
\name{diffusion_sim}
\alias{diffusion_sim}
\title{Drift Diffusion Simulation}
\usage{
diffusion_sim(
  v = 0,
  sv = 0,
  a = 2,
  w = 0.5,
  sw = 0,
  t0 = 0.2,
  st0 = 0,
  dt = 0.01,
  t_max = Inf
)
}
\arguments{
\item{v}{The average drift rate. Determines the rate the agent travels
towards evidence boundaries. When multiple trials are simulated, drift rates
are randomly sampled from a normal distribution with a mean of \code{v}.}

\item{sv}{The standard deviation of the drift rate. When multiple simulations
are run and \code{sv} > 0, drift rates change trial by trial based on random
sampling from a normal distribution.}

\item{a}{Response caution, represented by the distance between evidence
boundaries and the agent's starting point. Larger distances between the
starting point and decision threshold represent more caution as more evidence
is required for the agent to come to a decision.}

\item{w}{Average response bias. the avg ease of traveling to either boundary}

\item{sw}{Boundary/starting point deviation}

\item{t0}{Residual time: time lag between evidence accumulation and action}

\item{st0}{The mean of the residual time distribution}

\item{dt}{Delta t: the units by which time increments}

\item{t_max}{Maximum time: the max time the model will draw samples}
}
\value{
A \code{'data.table'} object containing the time and activation of the
agent until it reached a boundary condition and the simulation ended.
}
\description{
A helper function—used by the \code{\link[=drift_mod]{drift_mod()}} function—which runs a single
simulation of a drift diffusion process based on the specified parameters.
All model parameters should be numeric and of length 1, unless otherwise
specified.
}
\examples{
diffusion_sim()
}
