% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_mod.R
\name{drift_mod}
\alias{drift_mod}
\title{Toy Drift Diffusion Model}
\usage{
drift_mod(
  nsims = 1000,
  v = 0.5,
  sv = 0,
  a = 2,
  w = 0.5,
  sw = 0.9,
  t0 = 0,
  st0 = 0.2,
  dt = 0.01,
  t_max = Inf,
  plots = TRUE,
  show_legend = FALSE,
  ...
)
}
\arguments{
\item{nsims}{An integer (1000 by default). The number of simulations to run.}

\item{v}{The average drift rate. This parameter determines the rate the
agent will travel towards a boundary.}

\item{sv}{The standard deviation of the drift rate. When multiple simulations
are run, such as with the \code{'drift_mod()'} function, standard deviation > 1
allows drift rates to change from one simulation to the next based on
random sampling.}

\item{a}{Response caution, represented by the distance from the
starting point of the agent to either decision threshold. Larger distances
between the starting point and a threshold represent more caution because
more evidence is required for the agent to come to a decision.}

\item{w}{Response bias mean: the avg ease of traveling to either boundary}

\item{sw}{Boundary/starting point deviation}

\item{t0}{Residual time: time lag between evidence accumulation and action}

\item{st0}{The mean of the residual time distribution}

\item{dt}{Delta t: the units by which time increments}

\item{t_max}{Maximum time: the max time the model will draw samples}

\item{plots}{Logical (TRUE by default). Should plots be generated when the
model is run?}

\item{show_legend}{Logical (FALSE by default). Should the legend be shown
on the evidence plot?}

\item{...}{Optional arguments to be passed to \code{'theme_pcj()'} which is used to
style the generated plots.}
}
\value{
Returns a list containing:
\itemize{
\item The results of the simulation
\item A density plot of evidence accumulation over time
\item A plot of the rt distributions for one outcome over the other
\item The RT quintile-probability plot for each decision with points at the
0.1, 0.3, 0.5, 0.7, and 0.9 quintiles
}
}
\description{
A function to run and visualize the results of a drift diffusion model.
Serves as a wrapper for an internal function, \link{diffusion_sim}. Though this
function is useful for visualizing and understanding drift diffusion models
and the influence different parameters have on simulated results, models
generated by this function are produced from random data and are ultimately
not practically informative.
}
\examples{
\donttest{
# Run simulation with default parameters
results <- drift_mod()

# Access simulation data
results$sim_results

# Run without plots
results <- drift_mod(nsims = 100, plots = FALSE)

# Show legend on evidence plot
results <- drift_mod(nsims = 100, show_legend = TRUE)
}
}
\references{
Farrell, S., & Lewandowsky, S. (2018). Computational modeling of cognition and behavior. Cambridge University Press.
}
