% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ezddm}
\alias{ezddm}
\title{Fit a Diffusion Model}
\usage{
ezddm(
  prop_correct,
  rt_correct_variance_seconds,
  rt_correct_mean_seconds,
  n_trials,
  s = 1
)
}
\arguments{
\item{prop_correct}{Proportion correct.}

\item{rt_correct_variance_seconds}{The variance of the correct response times
(in seconds).}

\item{rt_correct_mean_seconds}{The mean of correct response times (in
seconds).}

\item{n_trials}{The number of trials.}

\item{s}{A scaling parameter (often set to 1.0—the default—or 0.1).}
}
\value{
A vector containing \code{'a'} (boundary),
\code{'v'} (drift rate), and \code{'ter'} (non-decision time) parameters.
}
\description{
This function will fit a diffusion model to subjects' aggregated data.
}
\examples{
ezddm(prop_correct = .802, rt_correct_variance_seconds = .112,
rt_correct_mean_seconds = .723, n_trials = 100)
}
