% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{files_info}
\alias{files_info}
\title{File Information Compiler}
\usage{
files_info(
  path = system.file("extdata", package = "pcjtools"),
  extension = "csv",
  ...
)
}
\arguments{
\item{path}{A filepath string that names a directory with files to read.}

\item{extension}{A string naming the extension of the files to import.}

\item{...}{Optional arguments passed to \code{list.files()}:
\code{all.files}, \code{recursive}, \code{ignore.case}, \code{include.dirs}, which all default
to FALSE.}
}
\value{
A list containing file information from files in the
directory specified in \code{'path'}. Includes the full filepath, the basename,
the extension, the last date the file was modified, and the age in days
since the file was last modified. This information can be stored in a vector
and is intended to be passed to \code{import_data()} to read in a list of data
files to a single \code{data.table} object.
}
\description{
A function to gather information about data files of the type specified in
the \code{'extension'} argument and located in the local directory specified in
\code{'path'}.
}
\examples{
files_info()
}
