% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{files_info}
\alias{files_info}
\title{a function to list information about data files from a local directory}
\usage{
files_info(path = NULL, extension, ...)
}
\arguments{
\item{path}{a filepath string naming a directory with files to read.}

\item{extension}{a string naming the extension of the files to import.}

\item{...}{optional arguments to be passed to the 'list.files()' function}
}
\value{
a data.table object containing file name strings from the specified
directory.
}
\description{
a function to list information about data files from a local directory
}
\examples{
files_info(path = NULL, extension = "csv")
}
