% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wienR.R
\name{fit_wienr}
\alias{fit_wienr}
\title{Fit a model}
\usage{
fit_wienr(
  rt,
  response,
  fit_sv = FALSE,
  fit_sw = FALSE,
  fit_st0 = FALSE,
  optim_control = list(),
  init_par = NULL,
  drift_index = NULL,
  bound_index = NULL,
  resid_index = NULL,
  sv_index = NULL,
  sw_index = NULL,
  st0_index = NULL
)
}
\arguments{
\item{rt}{Start}

\item{response}{Start}

\item{fit_sv}{Start}

\item{fit_sw}{Start}

\item{fit_st0}{Start}

\item{optim_control}{Start}

\item{init_par}{Start}

\item{drift_index}{Start}

\item{bound_index}{Start}

\item{resid_index}{Start}

\item{sv_index}{Start}

\item{sw_index}{Start}

\item{st0_index}{Start}
}
\value{
Something
}
\description{
Fit a model
}
\examples{
\dontrun{
fit_wienr()
}
}
