% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{A function to import jsPsych CSV data files downloaded from cognition.run}
\usage{
import_csv(file_dir = NULL, file_type = NULL)
}
\arguments{
\item{file_dir}{a string naming a directory filepath that contains
files to read.}

\item{file_type}{a string that gives the name of the
file's extension (e.g., "xlsx", "csv", etc.)}
}
\value{
a data.table object containing the read-in data
from the specified directory.
}
\description{
A function to import jsPsych CSV data files downloaded from cognition.run
}
\examples{
import_csv(file_type = "csv")
}
