% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Data Importer}
\usage{
import_data(x)
}
\arguments{
\item{x}{A string or list of strings containing the path of file(s) to
read in; can be the name of a \code{data.table} column in the form
\code{dt$filepaths}, which contains filepaths. All files in a list will be read
in and returned as a single \code{data.table} object.}
}
\value{
a data.table object containing the read-in data from the paths
specified in \code{x}.
}
\description{
A function to import data files from a local directory. Designed
to read a list of file path strings, principally, those stored in the
first column of the \code{data.table} produced by the \code{file_info()} function from
\code{'pcjtools'}.
}
\examples{
info <- files_info()
file <- import_data(x = info$filepath)

# alternatively:

file <- import_data(x = info[, "filepath"])
file <- import_data(x = info[, 1]) # Where '1' is the column with filepath(s)
}
