% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{a function to import data files from a local directory}
\usage{
import_data(path = NULL, extension = NULL, ...)
}
\arguments{
\item{path}{a filepath string naming a directory with files to read. All
files of the type named in extension will be read into a single data.table
object}

\item{extension}{a string naming a kind of file to import.}

\item{...}{optional arguments to be passed to list.files}
}
\value{
a data.table object containing the read-in data from the specified
directory.
}
\description{
a function to import data files from a local directory
}
\examples{
import_data(extension = "csv")
}
