% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R, R/theme_pcj.R
\name{install_atkinson}
\alias{install_atkinson}
\title{a function to install the 'Atkinson Hyperlegible' font, used by
'theme_pcj()'}
\usage{
install_atkinson()

install_atkinson()
}
\description{
a function to install the 'Atkinson Hyperlegible' font, used by
'theme_pcj()'

A function to install the Atkinson Hyperlegible font required for the
theme_pcj
}
\examples{
install_atkinson()
install_atkinson()
}
