% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_model.R
\name{memory_model}
\alias{memory_model}
\title{Run a memory model simulation}
\usage{
memory_model(
  seed = 20240709,
  n.letters = 12,
  n.trials = 500,
  updated.probs.sd = 0.1,
  max.forgetting = 0.6,
  forgetting.slope = 3,
  ...
)
}
\arguments{
\item{seed}{a single integer value used to set the state of randomly generated numbers to keep consistency when random numbers are sampled.}

\item{n.letters}{a single integer, the number of letters to be used in the memory model.}

\item{n.trials}{a single integer, the number of iterations the model should be run for.}

\item{updated.probs.sd}{a single integer, the standard deviation value to be used.}

\item{max.forgetting}{a single integer, the probability of forgetting any one item after you've learned it.}

\item{forgetting.slope}{a single integer, the rate you need to see an item to decrease the probability it will be forgotten.}

\item{...}{optional. Settings to pass to the theme() function.}
}
\value{
a matrix of learned memory strengths between letters, a matrix of the randomly generated probabilities used when drawing random samples of letters, and three graphs of the strengths between letters in memory the model's memory representation.
}
\description{
Run a memory model simulation
}
\examples{
memory_model(seed = 1234, n.letters = 12, n.trials = 200)
}
