% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{pcj_aesthetics}
\alias{pcj_aesthetics}
\title{Aesthetics to maximize legibility of graphs printed for academic posters.
This function is wrapped in theme_pcj and cannot be used directly.}
\usage{
pcj_aesthetics(base_size = 12, dark_text = "#000000")
}
\arguments{
\item{base_size}{an integer. The graph's default font size.}

\item{dark_text}{a quoted hex string, sets the color of the darkest text in a
plot. All text is based on various shades of the specified hex code.}
}
\value{
a plot configured with aesthetic settings specified by options set by
this function
}
\description{
Aesthetics to maximize legibility of graphs printed for academic posters.
This function is wrapped in theme_pcj and cannot be used directly.
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
  ggplot2::geom_point() +
  pcj_aesthetics()
}
