% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{pcj_graph_palettes}
\alias{pcj_graph_palettes}
\title{A collection of color palettes to be used in the generation of ggplot graphs.
This function is wrapped in theme_pcj and cannot be used directly.}
\usage{
pcj_graph_palettes(
  palette = "default",
  continuous = FALSE,
  .colors = pcj_colors,
  .palettes = pcj_palettes
)
}
\arguments{
\item{palette}{a string, including: "default", "neg_to_pos", "mono_printing",
"mono_blue", "mono_red", "mono_yellow", or "ualbany".}

\item{continuous}{logical. Whether the palette used in the graph should use a
discrete or continuous scale.}

\item{.colors}{loads colors defined in pcj_colors.}

\item{.palettes}{loads palettes defined in pcj_palettes.}
}
\value{
a plot object with the specified aesthetics rendered.
}
\description{
A collection of color palettes to be used in the generation of ggplot graphs.
This function is wrapped in theme_pcj and cannot be used directly.
}
\examples{
pcj_graph_palettes(palette = "default")
pcj_graph_palettes(palette = "mono_printing")
}
