% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcjtools-package.R
\docType{package}
\name{pcjtools-package}
\alias{pcjtools}
\alias{pcjtools-package}
\title{A suite of tools built to make data wrangling and analysis easier}
\description{
A collection of tools developed to make data analysis in R easier.
}
\author{
\strong{Maintainer}: Pierce Johnson \email{johnsonpierce.c@gmail.com} (\href{https://orcid.org/0009-0002-8517-2452}{ORCID})

}
\keyword{internal}
