% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcjtools-package.R
\docType{package}
\name{pcjtools-package}
\alias{pcjtools}
\alias{pcjtools-package}
\title{pcjtools: A set of tools developed for wrangling and analyzing data from jsPsych}
\description{
Convenience functions to make some common data wrangling and analysis of experimental data from jsPsych and Congition.run a bit easier.
}
\author{
\strong{Maintainer}: Pierce Johnson \email{johnsonpierce.c@gmail.com} (\href{https://orcid.org/0009-0002-8517-2452}{ORCID})

}
\keyword{internal}
