% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saver.R
\name{plot_saver}
\alias{plot_saver}
\title{Efficiently Save Plot(s) to File with Consistent Settings}
\usage{
plot_saver(plots, dir = ".", names = NULL, ...)
}
\arguments{
\item{plots}{A ggplot object or list of gglot objects to save to file}

\item{dir}{A string or character vector containing the path to the directory
where plot(s) will be saved.}

\item{names}{A name string, list of name strings that matches the length of
the \code{'plot'} list, or \code{NULL} (the default). Names to give to the plot files.
If a single name string and > 1 plot is provided, each plot will be
given that name followed by a numeric counter.}

\item{...}{Optional arguments to be passed to \code{'ggsave()'}.}
}
\value{
(Invisibly) returns a list of the plots saved to file in the
location given in \code{'dir'}.
}
\description{
Efficiently Save Plot(s) to File with Consistent Settings
}
\examples{
\donttest{
library(ggplot2)

g1 <- ggplot(
  data = mtcars,
  mapping = aes(x = mpg, y = wt, color = factor(cyl))
) +
  geom_point()

plot_saver(plots = g1, dir = tempdir())
}
}
