% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{q_wdm}
\alias{q_wdm}
\title{Title}
\usage{
q_wdm(p, response, ...)
}
\arguments{
\item{p}{Start}

\item{response}{Start}

\item{...}{Optional arguments to pass to the functions from the \code{'WienR'}
package.}
}
\value{
Something
}
\description{
Title
}
\examples{
\dontrun{
q_wdm()
}
}
