% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_fit.R
\name{qp_fit}
\alias{qp_fit}
\title{Title}
\usage{
qp_fit(
  rt,
  response,
  par = NULL,
  rt_p = c(0.1, 0.3, 0.5, 0.7, 0.9),
  drift_index = NULL,
  bound_index = NULL,
  resid_index = NULL,
  sv_index = NULL,
  sw_index = NULL,
  st0_index = NULL
)
}
\arguments{
\item{rt}{The RTs of decisions}

\item{response}{What the subject chose a trial}

\item{par}{Parameters}

\item{rt_p}{The percentiles to calculate from the RT distribution.}

\item{drift_index}{The index of the drift rate value.}

\item{bound_index}{The index of the boundary.}

\item{resid_index}{The index of residual}

\item{sv_index}{Trial-level variance of the draft rate}

\item{sw_index}{Variance in the starting position (I think)}

\item{st0_index}{Variance of the non-decision time}
}
\value{
Something
}
\description{
Title
}
\examples{
\dontrun{
qp_fit(rt = data_sub$rt,
response = data_sub$correct,
par = ddm_fit$par)
}
}
