% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{read_file}
\alias{read_file}
\title{File Reader}
\usage{
read_file(x)
}
\arguments{
\item{x}{A filepath in the form of a string.}
}
\value{
A \code{data.table} object containing data from a file supplied to
'read_data_list()'
}
\description{
A helper function used by \code{'read_file_list()'} to read individual data files.
}
\examples{
info <- files_info()
file <- import_data(x = info$filepath)
}
\keyword{internal}
