% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{read_file}
\alias{read_file}
\title{a helper function used by 'read_data_list()' to read individual data files}
\usage{
read_file(file)
}
\arguments{
\item{file}{a string or list of strings with file name paths}
}
\value{
a data.table object containing data from a file supplied to
'read_data_list()'
}
\description{
a helper function used by 'read_data_list()' to read individual data files
}
\examples{
data <- read_file(file)
}
