% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_file}
\alias{read_file}
\title{File Reader}
\usage{
read_file(x)
}
\arguments{
\item{x}{A filepath in the form of a string.}
}
\value{
A \code{data.table} object containing data from a file supplied to
\code{'read_data_list()'}.
}
\description{
A helper used by \code{'read_file_list()'} to read in individual data files.
}
\examples{
\dontrun{
info <- files_info()
file <- read_file(x = info$filepath)
}
}
\keyword{internal}
