% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_file}
\alias{read_file}
\title{File Reader}
\usage{
read_file(x)
}
\arguments{
\item{x}{A string giving the file path of a single file to read in.}
}
\value{
A \code{data.table} object containing data from a file supplied to
\code{'read_file_list()'}.
}
\description{
A helper used by \code{\link[=read_file_list]{read_file_list()}} to read in individual data files.
}
\keyword{internal}
