% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_file_list}
\alias{read_file_list}
\title{A Reader for Lists of Files}
\usage{
read_file_list(files)
}
\arguments{
\item{files}{A string or an object containing filepath string(s).}
}
\value{
A \code{data.table} object with concatenated data from all named files
supplied in \code{'file_list'}.
}
\description{
A wrapper to read in multiple files passed from the \code{'read_file()'} helper
and bind them together into a single object.
}
\examples{
\dontrun{
info <- files_info()
data <- read_file_list(files = info$filepath)
}
}
\keyword{internal}
