% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{read_file_list}
\alias{read_file_list}
\title{a function to read in data passed from the 'read_file()' helper function}
\usage{
read_file_list(files)
}
\arguments{
\item{files}{a string or an object containing strings with file path(s)}
}
\value{
a data.table object of concatenated data from all named files
supplied in 'file_list'
}
\description{
a function to read in data passed from the 'read_file()' helper function
}
\examples{
data <- read_file_list(files = files_info())
}
