% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_file_list}
\alias{read_file_list}
\title{A Reader for Lists of Files}
\usage{
read_file_list(
  files,
  parallel = FALSE,
  n_cores = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{files}{A string or an object containing filepath string(s).}

\item{parallel}{Logical; use parallel processing? Default is FALSE.}

\item{n_cores}{Integer; number of cores to use. Default uses all but one core.}

\item{show_progress}{Logical; show progress bar? Default is TRUE.}

\item{...}{
  Arguments passed on to \code{\link[data.table:rbindlist]{data.table::rbindlist}}
  \describe{
    \item{\code{l}}{ A list containing \code{data.table}, \code{data.frame} or \code{list} objects. \code{\dots} is the same but you pass the objects by name separately. }
    \item{\code{use.names}}{\code{TRUE} binds by matching column name, \code{FALSE} by position. \code{"check"} (default) warns if all items don't have the same names in the same order and then currently proceeds as if \code{use.names=FALSE} for backwards compatibility (\code{TRUE} in future); see news for v1.12.2.}
    \item{\code{fill}}{\code{TRUE} fills missing columns with NAs, or NULL for missing list columns. By default \code{FALSE}.}
    \item{\code{idcol}}{Creates a column in the result showing which list item those rows came from. \code{TRUE} names this column \code{".id"}. \code{idcol="file"} names this column \code{"file"}. If the input list has names, those names are the values placed in this id column, otherwise the values are an integer vector \code{1:length(l)}. See \code{examples}.}
    \item{\code{ignore.attr}}{Logical, default \code{FALSE}. When \code{TRUE}, allows binding columns with different attributes (e.g. class).}
  }}
}
\value{
A \code{data.table} object with concatenated data from all named files.
}
\description{
A wrapper to read in multiple files and bind them together into a single object.
}
\keyword{internal}
