% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{theme_pcj}
\alias{theme_pcj}
\title{a function to create and save consistent plots}
\usage{
theme_pcj(
  plot,
  base_size = 12,
  dark_text = "#000000",
  font = "Atkinson Hyperlegible",
  palette = "default",
  continuous = FALSE,
  plot_text = c("title", "subtitle", "xlab", "ylab", paste0("Rendered:",
    format(Sys.time(), "\%Y\%m\%d, \%H:\%M"))),
  alt_text = TRUE,
  save = FALSE,
  ...
)
}
\arguments{
\item{plot}{a ggplot object to be formatted}

\item{base_size}{an integer. The graph's default font size.}

\item{dark_text}{a quoted hex string. Sets the color of the darkest text in a
plot. All text is based on shades of the specified hex code.}

\item{font}{a quoted name of the font you would like the plot's text to be
printed in}

\item{palette}{a string. The name of the palette to be mapped to a variable,
including: "default", "neg_to_pos", "mono_printing", "mono_blue", "mono_red",
"mono_yellow", or "ualbany".}

\item{continuous}{logical. Is the variable continuous or discrete?}

\item{plot_text}{a named list of plot features and their text label}

\item{alt_text}{logical. Should a subtitle and caption be generated for the
plot?}

\item{save}{logical. Should the rendered plot be saved to the working
directory}

\item{...}{optional arguments to be passed to 'theme_pcj_aesthetics()',
'theme_pcj_palettes()', or 'theme_pcj_text()'}
}
\value{
a plot object with the specified aesthetics
}
\description{
a function to create and save consistent plots
}
\examples{
\dontrun{
theme_pcj(plot = plot)
}
}
