% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{theme_pcj}
\alias{theme_pcj}
\title{A theme developed to build consistent ggplot graphs}
\usage{
theme_pcj(
  ggplot_object,
  palette = "default",
  continuous = FALSE,
  base_size = 12,
  dark_text = "#000000",
  graph_text = list(title = "title", subtitle = "subtitle", xlab = "xlab", ylab = "ylab",
    caption = paste("Revised", Sys.time()))
)
}
\arguments{
\item{ggplot_object}{a ggplot object stored in a vector.}

\item{palette}{a string, including: "default", "neg_to_pos", "mono_printing",
"mono_blue", "mono_red", "mono_yellow", or "ualbany".}

\item{continuous}{logical. Whether the scale of the data is discrete or continuous.}

\item{base_size}{an integer. The graph's default font size.}

\item{dark_text}{a quoted hex string, sets the color of the darkest text in a plot. All text is based on various shades of the specified hex code.}

\item{graph_text}{a named list of plot labels and the text to fill them.}
}
\value{
a ggplot object following the specified settings
}
\description{
A theme developed to build consistent ggplot graphs
}
\examples{
g1 <- ggplot2::ggplot(
  data = mtcars,
  ggplot2::aes(x = mpg, y = wt, color = factor(cyl))
) +
  ggplot2::geom_point()
theme_pcj(ggplot_object = g1)
}
