% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pcj.R
\name{theme_pcj}
\alias{theme_pcj}
\title{A theme made to maximize legibility of text on graphs printed for academic posters}
\usage{
theme_pcj(base_size = 12, dark_text = "#000000", os_fonts = "win")
}
\arguments{
\item{base_size}{an integer, the size the font in a plot should default to.}

\item{dark_text}{a quoted hex string, sets the color of the darkest text in a plot.}

\item{os_fonts}{a string, including "all", "pdf", "postscript" and "win", indicating which fonts should be loaded into RStudio}
}
\value{
a plot configured with aesthetic settings specified by options set by this function
}
\description{
A theme made to maximize legibility of text on graphs printed for academic posters
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
ggplot2::geom_point() +
theme_pcj()
}
