% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{theme_pcj_aesthetics}
\alias{theme_pcj_aesthetics}
\title{a helper function defining theme elements to make consistent plots and graphs
with 'theme_pcj()'}
\usage{
theme_pcj_aesthetics(base_size = 12, dark_text = "#000000", ...)
}
\arguments{
\item{base_size}{an integer. The graph's default font size.}

\item{dark_text}{a quoted hex string. Sets the color of the darkest text in a
plot. All text is based on shades of the specified hex code.}

\item{...}{additional arguments passed to 'theme()'}
}
\value{
a plot configured with aesthetics reflecting the specified settings
}
\description{
a helper function defining theme elements to make consistent plots and graphs
with 'theme_pcj()'
}
\examples{
ggplot2::ggplot(data = mtcars, ggplot2::aes(x = wt, y = mpg, color = gear)) +
ggplot2::geom_point() +
theme_pcj_aesthetics()
}
