% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{theme_pcj_palettes}
\alias{theme_pcj_palettes}
\title{a helper function defining the palettes accessible to 'theme_pcj()'}
\usage{
theme_pcj_palettes(
  palette,
  continuous,
  .colors = pcj_colors,
  .palettes = pcj_palettes
)
}
\arguments{
\item{palette}{a string. The name of the palette to be mapped to a variable,
including: "default", "neg_to_pos", "mono_printing", "mono_blue", "mono_red",
"mono_yellow", or "ualbany".}

\item{continuous}{logical. Is the variable continuous or discrete?}

\item{.colors}{loads the colors defined within pcj_colors.}

\item{.palettes}{loads the palettes defined within pcj_palettes.}
}
\value{
a plot object with the specified aesthetics rendered.
}
\description{
a helper function defining the palettes accessible to 'theme_pcj()'
}
\examples{
\dontrun{
pcj_graph_palettes(palette = "default")
pcj_graph_palettes(palette = "mono_printing")
}
}
\keyword{internal}
