% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_pcj_palettes}
\alias{theme_pcj_palettes}
\title{Plot Color Palettes}
\usage{
theme_pcj_palettes(
  palette,
  continuous,
  .colors = pcj_colors,
  .palettes = pcj_palettes
)
}
\arguments{
\item{palette}{A string. The name of the palette to be mapped to a variable,
including: "default", "neg_to_pos", "mono_printing", "mono_blue", "mono_red",
"mono_yellow", or "ualbany".}

\item{continuous}{Logical. Is the variable continuous or discrete?}

\item{.colors}{Loads the colors defined within pcj_colors.}

\item{.palettes}{Loads the palettes defined within pcj_palettes.}
}
\value{
A plot object with the specified aesthetics rendered.
}
\description{
A helper function defining the palettes accessible to \code{'theme_pcj()'}.
}
\keyword{internal}
