% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_pcj_text}
\alias{theme_pcj_text}
\title{Plot Text Settings}
\usage{
theme_pcj_text(plot_text, alt_text)
}
\arguments{
\item{plot_text}{A named character vector where plot features are names and
the text to be printed in the plot are values (e.g.,
c(title = "Plot Title", etc.)).}

\item{alt_text}{Logical. Should a subtitle and caption be generated for the
plot?}
}
\value{
A ggplot object with the text declared in the function call.
}
\description{
A helper to define the text for various elements of a plot as part of the
\code{'theme_pcj()'} function.
}
\keyword{internal}
