% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_git_package.R
\name{update_git_package}
\alias{update_git_package}
\title{Update Packages from Github}
\usage{
update_git_package(repo = NULL, upgrade = "always", force = FALSE)
}
\arguments{
\item{repo}{A repository address in the form 'username/repo'
(e.g., 'johnsonpc2/pcjtools').}

\item{upgrade}{Should the package's dependencies be upgraded? One of
"default", "ask", "always", or "never"}

\item{force}{Force installation, even if the remote state has not changed
since the previous install.}
}
\value{
Invisibly updates packages from Github.
}
\description{
A thin wrapper for install_github() from the devtools package. Used to update
packages from Github repositories.
}
\examples{
\dontrun{
update_git_package(repo = 'johnsonpc2/pcjtools')
}
}
